extract datalayer iterators to separate file
